+++ Testing/SystemTests/tests/framework/CMakeLists.txt
@@ -137,8 +137,9 @@ set(SYSTEST_NAMES
     MaxEntTest.py
     MDNormBackgroundHYSPECTest.py
     MDNormCORELLITest.py
+    MDNormDEMANDTest.py
     MDNormHYSPECTest.py
-    MDNormMonochromaticSCDTest.py
+    MDNormWANDTest.py
     MDWorkspaceTests.py
     MonteCarloAbsorptionCorrTest.py
     MPIAlgorithmsExecuteTest.py
